SX1262: some initial listen before talk code.
Unfortunatley the SX1276 was much better at this. It's a true problem
and I still believe the chip somewhat has the feature to report that
it is in the process of receiving a packet, but it surely is not
documented.

I also tried if the instantaneous RSS could be useful, but it's totally
useless with weak signals. The radio can read correct messages while
the RSS remains at incredibly low values, assuming my RSS calculation
is correct, but seems quite simple as documented in the datasheet.